start cleaning up the different NSE headers this is part of the way towards letting the approaches coexist
