+++ nse_solver/Make.package
@@ -1,5 +1,6 @@
 ifeq ($(USE_NSE_NET), TRUE)
      CEXE_headers += nse_solver.H
+     CEXE_headers += nse_check.H
      CEXE_headers += nse_index.H
 
      CEXE_sources += nse_index.cpp
