+++ nse_solver/README.md
@@ -1,0 +1,6 @@
+# Self-Consistent NSE Solver
+
+This is a nuclear statistical equilibrium solver (NSE) that works with
+any network and finds the NSE state of the nuclei known to the network
+given an input temperature, density, and electron fraction, Y_e.
+
